requirements.txt
absl-py==0.15.0
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
ankh==1.10.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
autopaths==1.6.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
biopython==1.81
bleach==6.0.0
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clang==5.0
click==8.1.6
cmake==3.26.4
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
cytoolz==0.12.2
datasets==2.13.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dgl-cu111==0.6.1
diamond4py==0.0.5
dill==0.3.6
editdistance==0.6.2
exceptiongroup==1.1.1
executing==1.2.0
fasta==2.3.2
fastjsonschema==2.17.1
filelock==3.12.2
flatbuffers==1.12
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.1.0
huggingface-hub==0.15.1
Hydra==2.5
hydra-core==1.3.2
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.23.3
ipython==8.12.2
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lit==16.0.6
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
munch==4.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.2.3
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
peft==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
plumbing==2.11.2
polars==0.18.9
profeatpy==0.0.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfastx==1.1.0
Pygments==2.15.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.1.99
setuptools==67.7.2
sh==2.0.6
six==1.15.0
sniffio==1.3.0
Sophia-Optimizer==0.2.1
soupsieve==2.4.1
speechbrain==0.5.15
stack-data==0.6.2
sympy==1.12
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
typing-extensions==3.10.0.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wget==3.2
wheel==0.40.0
wrapt==1.12.1
xgboost==1.7.6
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
